requirements.txt
# requirements.txt 
langchain 
langgraph 
langchain_groq
langchain_core 
langchain_huggingface
langchain-community
faiss-cpu 
sentence-transformers 
pandas
python_dotenv